% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgurobi.R
\docType{package}
\name{rapr}
\alias{rapr}
\alias{rapr-package}
\title{rgurobi: Solve optimisation problems using Gurobi}
\description{
Solve optimisation problems using the commerical 'Gurobi' software package (obtained from <http://www.gurobi.com/>). This package provides similar functionality to the official gurobi R package (see <http://www.gurobi.com/products/modeling-languages/r>). However, this package allows users to retrive multiple solutions after solving a problem. Note that this package does not support quadratic expressions.
}
\details{
Type \code{vignette('rgurobi')} for a tutorial on how to use this package.
}

